@@ -2,11 +2,12 @@
 @using Microsoft.AspNetCore.Hosting;
 @using Microsoft.AspNetCore.Localization
 @using Microsoft.AspNetCore.Mvc.Localization
+@using Microsoft.Extensions.Hosting;
 @using Microsoft.Extensions.Options
 @inject IHtmlLocalizer<SharedResource> SharedLocalizer;
 @inject Microsoft.AspNetCore.Http.IHttpContextAccessor HttpContextAccessor
 @inject IOptions<RequestLocalizationOptions> LocOptions
-@inject IHostingEnvironment env
+@inject IWebHostEnvironment env
 
 @{
     var requestCulture = Context.Features.Get<IRequestCultureFeature>();
